+++ src/backend/src/main/java/com/bill_split/app/data/User.java
@@ -13,7 +13,7 @@ public class User {
     private String email;
     
     @Column(name = "total_cost")
-    private Long total_cost = 0L;
+    private BigDecimal total_cost = 0;
     
     public Long getId() {
         return id;
@@ -31,11 +31,11 @@ public class User {
         this.email = email;
     }
     
-    public Long getTotalCost() {
+    public BigDecimal getTotalCost() {
         return total_cost;
     }
     
-    public void setTotalCost(Long total_cost) {
+    public void setTotalCost(BigDecimal total_cost) {
         this.total_cost = total_cost;
     }
 }
